Rename package to clipboard
