+++ clipboard/clipboard.go
@@ -1,4 +1,4 @@
-package system
+package clipboard
 
 import (
 	"context"
