+++ clipboard/clipboard_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package system
+package clipboard
 
 import (
 	"context"
